Add component for SidebarTabs
